moved network requester into gateway
