+++ Cargo.toml
@@ -133,7 +133,6 @@ members = [
     "sdk/rust/nym-sdk",
     "service-providers/common",
     "service-providers/ip-packet-router",
-    "service-providers/network-requester",
     "sqlx-pool-guard",
     "tools/echo-server",
     "tools/internal/contract-state-importer/importer-cli",
@@ -169,7 +168,6 @@ default-members = [
     "nym-validator-rewarder",
     "nyx-chain-watcher",
     "service-providers/ip-packet-router",
-    "service-providers/network-requester",
     "tools/nymvisor",
 ]
