+++ gateway/Cargo.toml
@@ -57,7 +57,6 @@ nym-gateway-stats-storage = { path = "../common/gateway-stats-storage" }
 nym-gateway-requests = { path = "../common/gateway-requests" }
 nym-mixnet-client = { path = "../common/client-libs/mixnet-client" }
 nym-network-defaults = { path = "../common/network-defaults" }
-nym-network-requester = { path = "../service-providers/network-requester" }
 nym-sdk = { path = "../sdk/rust/nym-sdk" }
 nym-sphinx = { path = "../common/nymsphinx" }
 nym-statistics-common = { path = "../common/statistics" }
@@ -71,13 +70,21 @@ nym-wireguard = { path = "../common/wireguard" }
 nym-wireguard-private-metadata-server = { path = "../common/wireguard-private-metadata/server" }
 nym-wireguard-types = { path = "../common/wireguard-types", default-features = false }
 
-nym-authenticator-requests = { path = "../common/authenticator-requests" }
 nym-client-core = { path = "../common/client-core", features = ["cli"] }
 nym-id = { path = "../common/nym-id" }
 nym-service-provider-requests-common = { path = "../common/service-provider-requests-common" }
 
-
 defguard_wireguard_rs = { workspace = true }
+
+nym-service-providers-common = { path = "../service-providers/common" }
+
+# authenticator specific
+nym-authenticator-requests = { path = "../common/authenticator-requests" }
+
+# network-requester specific
+nym-socks5-proxy-helpers = { path = "../common/socks5/proxy-helpers" }
+nym-socks5-requests = { path = "../common/socks5/requests" }
+nym-exit-policy = { path = "../common/exit-policy", features = ["client"] }
 
 [dev-dependencies]
 nym-gateway-storage = { path = "../common/gateway-storage", features = ["mock"] }
