+++ gateway/src/lib.rs
@@ -11,4 +11,6 @@ pub mod node;
 pub use error::GatewayError;
 pub use node::GatewayTasksBuilder;
 
+pub use node::internal_service_providers as service_providers;
 pub use node::internal_service_providers::authenticator as nym_authenticator;
+pub use node::internal_service_providers::network_requester as nym_network_requester;
