+++ gateway/src/node/internal_service_providers/network_requester/allowed.list.sample
