+++ gateway/src/node/internal_service_providers/network_requester/config/persistence.rs
