+++ gateway/src/node/internal_service_providers/network_requester/mod.rs
@@ -1,2 +1,13 @@
 // Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
+
+pub mod config;
+pub mod core;
+pub mod error;
+mod reply;
+pub mod request_filter;
+mod socks5;
+
+pub use config::Config;
+pub use core::{NRServiceProvider, NRServiceProviderBuilder, OnStartData};
+pub use request_filter::RequestFilter;
