+++ gateway/src/node/mod.rs
@@ -4,8 +4,11 @@
 use crate::config::Config;
 use crate::error::GatewayError;
 use crate::node::client_handling::websocket;
+use crate::node::internal_service_providers::authenticator::Authenticator;
+use crate::node::internal_service_providers::authenticator::{self};
+use crate::node::internal_service_providers::network_requester::{self, NRServiceProviderBuilder};
 use crate::node::internal_service_providers::{
-    authenticator, ExitServiceProviders, ServiceProviderBeingBuilt, SpMessageRouterBuilder,
+    ExitServiceProviders, ServiceProviderBeingBuilt, SpMessageRouterBuilder,
 };
 use crate::node::stale_data_cleaner::StaleMessagesCleaner;
 use futures::channel::oneshot;
@@ -17,7 +20,6 @@ use nym_crypto::asymmetric::ed25519;
 use nym_ip_packet_router::IpPacketRouter;
 use nym_mixnet_client::forwarder::MixForwardingSender;
 use nym_network_defaults::NymNetworkDetails;
-use nym_network_requester::NRServiceProviderBuilder;
 use nym_node_metrics::events::MetricEventsSender;
 use nym_node_metrics::NymNodeMetrics;
 use nym_task::{ShutdownToken, TaskClient};
@@ -33,10 +35,9 @@ use tracing::*;
 use zeroize::Zeroizing;
 
 pub(crate) mod client_handling;
-pub(crate) mod internal_service_providers;
+pub mod internal_service_providers;
 mod stale_data_cleaner;
 
-use crate::node::internal_service_providers::authenticator::Authenticator;
 pub use client_handling::active_clients::ActiveClientsStore;
 pub use nym_gateway_stats_storage::PersistentStatsStorage;
 pub use nym_gateway_storage::{
@@ -48,7 +49,7 @@ pub use nym_sdk::{NymApiTopologyProvider, NymApiTopologyProviderConfig, UserAgen
 
 #[derive(Debug, Clone)]
 pub struct LocalNetworkRequesterOpts {
-    pub config: nym_network_requester::Config,
+    pub config: network_requester::Config,
 
     pub custom_mixnet_path: Option<PathBuf>,
 }
