+++ service-providers/ip-packet-router/Cargo.toml
@@ -25,7 +25,6 @@ nym-exit-policy = { path = "../../common/exit-policy" }
 nym-id = { path = "../../common/nym-id" }
 nym-ip-packet-requests = { path = "../../common/ip-packet-requests" }
 nym-network-defaults = { path = "../../common/network-defaults" }
-nym-network-requester = { path = "../network-requester" }
 nym-sdk = { path = "../../sdk/rust/nym-sdk" }
 nym-service-provider-requests-common = { path = "../../common/service-provider-requests-common" }
 nym-service-providers-common = { path = "../common" }
