--- service-providers/network-requester/Cargo.toml
@@ -1,39 +1,0 @@
-# Copyright 2020 - Nym Technologies SA <contact@nymtech.net>
-# SPDX-License-Identifier: GPL-3.0-only
-
-[package]
-name = "nym-network-requester"
-license = "GPL-3.0"
-version = "1.1.62"
-authors.workspace = true
-edition.workspace = true
-rust-version = "1.70"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = { workspace = true }
-async-trait = { workspace = true }
-futures = { workspace = true }
-humantime-serde = { workspace = true }
-tracing = { workspace = true }
-reqwest = { workspace = true, features = ["json"] }
-serde = { workspace = true, features = ["derive"] }
-thiserror = { workspace = true }
-tokio = { workspace = true, features = ["net", "rt-multi-thread", "macros"] }
-url = { workspace = true }
-
-# internal
-nym-bin-common = { path = "../../common/bin-common" }
-nym-client-core = { path = "../../common/client-core", features = ["fs-gateways-storage", "fs-surb-storage"] }
-nym-config = { path = "../../common/config" }
-nym-network-defaults = { path = "../../common/network-defaults" }
-nym-sdk = { path = "../../sdk/rust/nym-sdk" }
-nym-service-providers-common = { path = "../common" }
-nym-socks5-proxy-helpers = { path = "../../common/socks5/proxy-helpers" }
-nym-socks5-requests = { path = "../../common/socks5/requests" }
-nym-sphinx = { path = "../../common/nymsphinx" }
-nym-task = { path = "../../common/task" }
-nym-exit-policy = { path = "../../common/exit-policy", features = ["client"] }
-nym-id = { path = "../../common/nym-id" }
-
